requirements.txt
requests
confluent-kafka
python-dotenv
cassandra-driver
